+++ CHANGELOG.md
@@ -51,6 +51,7 @@ Bug fixes:
 * Fixed `Symbol#to_proc` to create proc with nil `source_location` (#1663).
 * Make `GC.start` work with keyword arguments.
 * Fixed `Kernel#clone` for `nil`, `true`, `false`, `Integer`, and `Symbol`.
+* Make top-level methods available in `Context#getBindings()` (#1838).
 
 Compatibility: